python_demo/advanced_topics/ai_frameworks/projects/project1_chatbot/requirements.txt
# Core dependencies
langchain==0.1.0
streamlit==1.29.0
openai==1.3.0

# Optional: for local models
# transformers==4.35.0
# torch==2.1.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.0

# Memory and storage
chromadb==0.4.15

# UI enhancements
streamlit-chat==0.1.1
plotly==5.17.0

# Development and testing
pytest==7.4.0
black==23.9.0
flake8==6.1.0